Added Population
This includes the general population not included in the subpopulation which are subject to diseases.

Updated plots, and added progressbars along with alerts. The plots now includes a graphical setting where user can pick colors.

Known Bugs: It is only the first set of plots that are shown as default.