Updated data
The data has been updated and comes with recalibrated weights. Prior to this update Incident and Prevalent patient were considered as a whole population in this calculation.

Each patient group is now treated as two seperate populations. This means that all values prior to this update for type == 0 is now obsolete, and needs to be recalculated.